# Task 02/swagger.yaml
openapi: 3.0.0
info:
  title: Portfolio API
  version: 1.0.0
  description: A robust, stateless RESTful backend API.
paths:
  /items:
    get:
      summary: Retrieve a list of items
      responses:
        '200':
          description: A JSON array of items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
    post:
      summary: Create a new item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Item created successfully
        '400':
          description: Bad Request (Validation failed)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /items/{id}:
    put:
      summary: Update an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Item updated successfully
        '400':
          description: Bad request (Validation failed)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      summary: Delete an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer